Delft3D-Flow reader: Maked coordinates are mapped
Masked coordinates stop opendrift from calculating the pixel size of
the domain.

This creates a "phantom" array with the same size of the grid holding
the unknown positions of the masked coordinates.

It prescribe these positions at the same indices of the original
array but over a regular array of the same size with the mean dx,
dy of the original, placed southeast of the original array.

This solves the pixel size problem so the reader loads. The info of
the reader is however wrong.

Uppon loading the reader Opendrift shows as the domain coverage
roughly the new "phantom" grid. From reading the code it seems that
is using the difference of the 1st two nodes in each dimension and
the number of rows and columns to calculate the extents of the domain.
When the 1rst 2 grid points in each direction are masked, it picks up
the points of the "phantom" grid.

It is not clear at this point (asked the lead software programmer)
if this impacts the ability of the model to actually use this reader
when particle is inside the actual domain, and in this case, outside
of the corners indicated by the reader.

Refs #5333 and fixes #5449