Delft3D-Flow reader: 3D output of get_variables
Corrects the output of 4D variables at a single timestep so that the
return of `_interpolate_profiles()` is converted into a (Z, Y, X)
cube.

4D variables when the `z` argument to `get_variables()` is None,
result in the interpolation of the profiles to the minimum z-level of
the top sigma layer, resulting in an approximation of the surface
values for the variable with the shape (Y, X).

Refs #5333 closes #5464