Delft3d-Flow reader: masked coord to nearest valid
Maps masked coordinates to nearest valid coordinates.

This is a memory intensive process that runs at the initialization
of the reader. It may give trouble it the grids are too large
horizontally since it needs a M x U grid to calculate the distances
between the total masked points (M) and the total unmasked points
(U).

Closes #5465