Delft3d-Flow reader: solves native d3d landmask
Changes to `_get_mask()` allows it to be used as `_get_data_method()`
when `land_binary_mask` is requested.

This patch, however, comments `land_binary_mask` from the supplied
variables since the use of the mask as land boundary doesn't yield
the necessary resolution and, when avaiable, preceds any additional
non-automatic land mask such as an external shapefile.

Closes #5466